requirements.txt
arrow==0.15.5
asgiref==3.2.3
attrs==19.3.0
autobahn==19.11.1
Automat==0.8.0
bleach==3.1.0
bleach-whitelist==0.0.10
blessed==1.16.1
boto3==1.10.39
botocore==1.13.39
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==3.0
django-crispy-forms==1.8.1
django-markdownify==0.8.0
django-markup==1.4.1
django-picklefield==2.0
django-q==1.0.2
django-storages==1.8
djangorestframework==3.11.0
docutils==0.15.2
enum-compat==0.0.3
feedparser==5.2.1
gunicorn==20.0.4
h11==0.9.0
html5lib==1.0.1
httptools==0.0.13
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
Markdown==3.1.1
mysqlclient==1.4.6
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
python-creole==1.3.2
python-dateutil==2.8.0
pytz==2019.3
redis==3.3.11
requests==2.22.0
s3transfer==0.2.1
service-identity==18.1.0
six==1.13.0
smartypants==2.0.1
sqlparse==0.3.0
textile==3.0.4
Twisted==19.10.0
txaio==18.8.1
urllib3==1.25.7
wcwidth==0.1.8
webencodings==0.5.1
websockets==8.1
zope.interface==4.7.1
